Validate in differente class
