+++ src/main/java/org/unir/javabeans/Suma.java
@@ -136,7 +136,7 @@ public class Suma {
         return num1 + num2;
     }
 
-    /*
+    /**
      * Suma dos números enteros.
      *
      * Este metodo realiza la suma de dos números enteros, pero imprime un mensaje
@@ -161,7 +161,7 @@ public class Suma {
         return num3 + num4;
     }
 
-    /*
+    /**
      * Suma tres números reales.
      *
      * Este metodo realiza la suma acumulativa de tres números reales, pero imprime
@@ -203,40 +203,5 @@ public class Suma {
         return acumulator;
     }
 
-    public Double validateIsDouble() {
-        Scanner sc = new Scanner(System.in);
-        Double num = null;
-        boolean valido = false;
-
-        while (!valido) {
-            System.out.print("Introduzca un numero real: ");
-            if (sc.hasNextDouble()) {
-                num = sc.nextDouble();
-                valido = true;
-            } else {
-                System.out.println("Entrada invalida, introduzca un numero real: ");
-                sc.next();
-            }
-        }
-        return num;
-    }
-
-    public Integer validateIsInteger() {
-        Scanner sc = new Scanner(System.in);
-        Integer num = null;
-        boolean valido = false;
-
-        while (!valido) {
-            System.out.print("Introduzca un numero entero: ");
-            if (sc.hasNextInt()) {
-                num = sc.nextInt();
-                valido = true;
-            } else {
-                System.out.println("Entrada invalida, introduzca un numero entero: ");
-                sc.next();
-            }
-        }
-        return num;
-    }
 
 }
